Refactor: Move HomeScreen components to attendanceScreen package
This commit moves the following composables from the `homeScreen` package to a new `attendanceScreen` subpackage:
- `AddSubjectButton.kt`
- `SubjectCard.kt`
- `TopBar.kt`

The `HomeScreen.kt` has been updated to use `NavHost` and now navigates to the `AttendanceScreen` which contains the previously mentioned composables.
`MainActivity.kt` has been updated to reflect this change in navigation, now navigating to `AttendanceScreen` instead of `HomeScreen` when returning from `SubjectDetailScreen`.