Made TimeTableSlot subjectId nullable and added functionality to add slots
The `subjectId` in the `TimeTable` data class is now nullable.
Tapping on an empty space in the `TimeTableDay` composable now adds a new one-hour `TimeTableSlot` at the tapped location. The `subjectId` for these new slots will be null.
The vertical positioning of `TimeTableSlot` has been adjusted to align with the center of the hour.
Constants `MILLIS_IN_HOUR` and `MILLIS_IN_DAY` have been introduced.